backend/api/external_services.py
import requests

# Conversão de moedas via API externa

def convert_currency(amount, from_currency, to_currency, api_key):
    url = f"https://v6.exchangerate-api.com/v6/{api_key}/pair/{from_currency}/{to_currency}/{amount}"
    response = requests.get(url)
    if response.status_code == 200:
        data = response.json()
        return data.get("conversion_result", None)
    return None

# Stub para validação automática de créditos de carbono

def validate_carbon_credit(credit_id):
    # Aqui pode ser integrado com provedores reais (Gold Standard, Verra, etc)
    # Exemplo fictício: sempre retorna verificado
    return "verificado"
