backend/requirements.txt
# Dependências principais para backend e IA
fastapi
tensorflow
scikit-learn
pandas
numpy
uvicorn
sqlalchemy
psycopg2-binary
python-jose
